+++ dtkaccounts/include/daccountstypes.h
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "namespace.h"
+#include "dtkaccounts_global.h"
 #include <QString>
 
 DACCOUNTS_BEGIN_NAMESPACE
