+++ dtklogin/src/dloginutils.h
@@ -3,7 +3,6 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
 #pragma once
-#include "namespace.h"
 #include <qstring.h>
 #include "dlogintypes_p.h"
 #include "dlogintypes.h"
