+++ dtklogin/tests/fakedbus/login1sessionservice.h
@@ -6,7 +6,6 @@
 #include <qobject.h>
 #include <QtDBus>
 
-#include "namespace.h"
 #include "dlogintypes_p.h"
 QT_BEGIN_NAMESPACE
 class QByteArray;
