+++ dtklogin/tests/ut_dloginseat.cpp
@@ -9,7 +9,6 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 #include "ddbusinterface.h"
-#include "namespace.h"
 
 DLOGIN_USE_NAMESPACE
