+++ dtklogin/tests/ut_dloginuser.cpp
@@ -9,7 +9,6 @@
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
 #include "ddbusinterface.h"
-#include "namespace.h"
 #include "dloginutils.h"
 
 DLOGIN_USE_NAMESPACE
