+++ dtkpower/include/dpowermanager.h
@@ -5,7 +5,7 @@
 #pragma once
 
 #include <qobject.h>
-#include "namespace.h"
+#include "dtkpower_global.h"
 
 DPOWER_BEGIN_NAMESPACE
 class DPowerManagerPrivate;
