+++ dtkpower/include/dtkpower_global.h
